+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.19)
+
+set(CMAKE_CXX_STANDARD 20)
+
+add_subdirectory(apps)
+#add_subdirectory(lib)
