+++ lib/kb.cpp
